fix: PR requirements
