+++ libs/vre/shared/app-representations/src/index.ts
@@ -13,6 +13,6 @@ export * from './lib/video-preview/video-preview.component';
 export * from './lib/representation.service';
 export * from './lib/value-operation-event.service';
 export * from './lib/file-representation';
-export * from './lib/region.interface';
+export * from './lib/region';
 export * from './lib/region.service';
 export * from './lib/get-file-value';
