+++ libs/vre/shared/app-representations/src/lib/file-representation.ts
@@ -6,7 +6,7 @@ import {
   ReadMovingImageFileValue,
   ReadStillImageFileValue,
 } from '@dasch-swiss/dsp-js';
-import { Region } from './region.interface';
+import { Region } from './region';
 
 /**
  * represents a file value including its annotations.
