+++ libs/vre/shared/app-resource-page/src/lib/resource.component.ts
@@ -1,8 +1,11 @@
 import { Component, Input, OnChanges } from '@angular/core';
 import { Title } from '@angular/platform-browser';
 import { Router } from '@angular/router';
-import { DspResource } from '@dasch-swiss/vre/shared/app-common';
-import { RegionService, ValueOperationEventService, getFileValue } from '@dasch-swiss/vre/shared/app-representations';
+import { Constants, CountQueryResponse, ReadFileValue } from '@dasch-swiss/dsp-js';
+import { DspCompoundPosition, DspResource } from '@dasch-swiss/vre/shared/app-common';
+import { getFileValue, RegionService, ValueOperationEventService } from '@dasch-swiss/vre/shared/app-representations';
+import { IncomingService } from '@dasch-swiss/vre/shared/app-resource-properties';
+import { take } from 'rxjs/operators';
 import { CompoundService } from './compound/compound.service';
 import { ResourcePageService } from './resource-page.service';
 
@@ -14,14 +17,17 @@ import { ResourcePageService } from './resource-page.service';
 })
 export class ResourceComponent implements OnChanges {
   @Input({ required: true }) resource!: DspResource;
-
+  representationsToDisplay!: ReadFileValue;
+  isCompoundNavigation: boolean | null = null;
   showRestrictedMessage = true;
   resourceIsObjectWithoutRepresentation!: boolean;
 
   constructor(
     private _router: Router,
     private _titleService: Title,
-    public resourcePageService: ResourcePageService
+    private _incomingService: IncomingService,
+    private _compoundService: CompoundService,
+    private _regionService: RegionService
   ) {
     this._router.events.subscribe(() => {
       this._titleService.setTitle('Resource view');
@@ -32,6 +38,52 @@ export class ResourceComponent implements OnChanges {
     this.showRestrictedMessage = true;
     this.resourceIsObjectWithoutRepresentation = getFileValue(this.resource) === null;
 
-    this.resourcePageService.onInit(this.resource);
+    this.onInit(this.resource);
+  }
+
+  onInit(resource: DspResource) {
+    if (this._isObjectWithoutRepresentation(resource)) {
+      this._checkForCompoundNavigation(resource);
+      return;
+    }
+
+    this.isCompoundNavigation = false;
+    this.representationsToDisplay = getFileValue(resource)!;
+
+    if (this._isImageWithRegions(resource)) {
+      this._regionService.onInit(resource);
+    }
+  }
+
+  pageIsLoaded() {
+    return (
+      this.isCompoundNavigation === false ||
+      (this.isCompoundNavigation === true && this._compoundService.incomingResource !== undefined)
+    );
+  }
+
+  private _isImageWithRegions(resource: DspResource) {
+    return resource.res.properties[Constants.HasStillImageFileValue] !== undefined;
+  }
+
+  private _isObjectWithoutRepresentation(resource: DspResource) {
+    return getFileValue(resource) === null;
+  }
+
+  private _checkForCompoundNavigation(resource: DspResource) {
+    this._incomingService
+      .getStillImageRepresentationsForCompoundResource(resource.res.id, 0, true)
+      .pipe(take(1))
+      .subscribe(countQuery => {
+        const countQuery_ = countQuery as CountQueryResponse;
+
+        if (countQuery_.numberOfResults > 0) {
+          this.isCompoundNavigation = true;
+          this._compoundService.onInit(new DspCompoundPosition(countQuery_.numberOfResults), this.resource);
+          return;
+        }
+
+        this.isCompoundNavigation = false;
+      });
   }
 }
